+++ eras/conway/impl/src/Cardano/Ledger/Conway/Rules/Pool.hs
@@ -1,6 +1,4 @@
 {-# LANGUAGE DataKinds #-}
-{-# LANGUAGE DeriveAnyClass #-}
-{-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE EmptyDataDeriving #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE ScopedTypeVariables #-}
@@ -11,17 +9,27 @@
 
 module Cardano.Ledger.Conway.Rules.Pool (
   ConwayPOOL,
-  ConwayPoolEvent (..),
-  ConwayPoolPredFailure (..),
 )
 where
 
 import Cardano.Ledger.BaseTypes (ShelleyBase)
-import Cardano.Ledger.CertState (PState)
+import Cardano.Ledger.CertState (PState (..))
 import Cardano.Ledger.Conway.Era (ConwayPOOL)
-import Cardano.Ledger.Core (Era (EraCrypto), EraRule)
-import Cardano.Ledger.Shelley.API (PoolCert, PoolEnv)
-import Control.DeepSeq (NFData)
+import Cardano.Ledger.Core (
+  EraCrypto,
+  EraPParams,
+  EraRule,
+  PoolCert,
+ )
+import Cardano.Ledger.Shelley.API (
+  PoolEnv,
+ )
+import Cardano.Ledger.Shelley.Rules (
+  PoolEvent,
+  ShelleyPoolPredFailure,
+  poolCertTransition,
+ )
+import Cardano.Ledger.Shelley.TxCert (ShelleyEraTxCert)
 import Control.State.Transition (
   BaseM,
   Environment,
@@ -32,21 +40,16 @@ import Control.State.Transition (
   State,
   transitionRules,
  )
-import GHC.Generics (Generic)
-import NoThunks.Class (NoThunks (..))
-
-data ConwayPoolPredFailure era
-  = ConwayPoolPredFailure
-  deriving (Show, Eq, Generic, NoThunks, NFData)
-
-newtype ConwayPoolEvent era = PoolEvent (Event (EraRule "POOL" era))
+import Control.State.Transition.Extended (TRC (TRC), TransitionRule, judgmentContext)
 
 instance
-  ( Era era
+  forall era.
+  ( EraPParams era
   , State (EraRule "POOL" era) ~ PState era
   , Signal (EraRule "POOL" era) ~ PoolCert (EraCrypto era)
   , Environment (EraRule "POOL" era) ~ PoolEnv era
   , EraRule "POOL" era ~ ConwayPOOL era
+  , ShelleyEraTxCert era
   ) =>
   STS (ConwayPOOL era)
   where
@@ -54,7 +57,17 @@ instance
   type Signal (ConwayPOOL era) = PoolCert (EraCrypto era)
   type Environment (ConwayPOOL era) = PoolEnv era
   type BaseM (ConwayPOOL era) = ShelleyBase
-  type PredicateFailure (ConwayPOOL era) = ConwayPoolPredFailure era
-  type Event (ConwayPOOL era) = ConwayPoolEvent era
+  type PredicateFailure (ConwayPOOL era) = ShelleyPoolPredFailure era
+  type Event (ConwayPOOL era) = PoolEvent era
 
-  transitionRules = undefined -- TODO
+  transitionRules = [conwayPoolTransition]
+
+conwayPoolTransition ::
+  ( EraPParams era
+  , ShelleyEraTxCert era
+  , EraRule "POOL" era ~ ConwayPOOL era
+  ) =>
+  TransitionRule (ConwayPOOL era)
+conwayPoolTransition = do
+  TRC (penv, pState, c) <- judgmentContext
+  poolCertTransition penv pState c